When generating C projects in NRPy; e.g., inside the project directory curviwavetoy/SinhCylindrical/, you'll see stuff like this:
... while the Makefile for the project is in curviwavetoy/ . To ensure that C codes in subdirectories look for headers relative to the root directory, there's some clunky logic in c_function.py:

                if "<" in inc:
                    complete_func += f"#include {inc}\n"
                else:
                    if par.parval_from_str("Infrastructure") == "BHaH":
                        if any(
                            x in inc
                            for x in [
                                "BHaH_defines.h",
                                "BHaH_function_prototypes.h",
                                "simd_intrinsics.h",
                                "cuda_intrinsics.h",
                            ]
                        ):
                            inc = os.path.join(rel_path_to_root_directory, inc)
                    complete_func += f'#include "{inc}"\n'
This prepends "../" in front of the header files listed above.

This is not a good design decision: in BH@H, we have additional header files, for interpolation and other purposes.

Instead of continuing to add to this clunky logic, I've decided to remove it in favor of the best practice for handling includes in subdirectories: adding -I. to all default cflags. This tells the C preprocessor to search the path "." [root directory] when processing #include's in subdirectories. The net effect is, e.g., the original example above changes to:

Also the above logic in c_function.py becomes:

                if "<" in inc:
                    complete_func += f"#include {inc}\n"
                else:
                    complete_func += f'#include "{inc}"\n'

Much cleaner!